moved Exceptions to exceptions package, removed odd tests
