+++ tests/src/androidTest/java/com/virgilsecurity/ethree/ExampleInstrumentedTest.kt
@@ -2,11 +2,9 @@ package com.virgilsecurity.ethree
 
 import android.support.test.InstrumentationRegistry
 import android.support.test.runner.AndroidJUnit4
-
+import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
-
-import org.junit.Assert.*
 
 /**
  * Instrumented test, which will execute on an Android device.
