+++ testscoroutines/src/androidTest/java/com/android/virgilsecurity/testscoroutines/ExampleInstrumentedTest.java
@@ -40,7 +40,7 @@ import android.support.test.runner.AndroidJUnit4;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 /**
  * Instrumented test, which will execute on an Android device.
